fix win64 build
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@272390 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 52e5a28e8bdf8ecfc06e7de206b673d96c2d95bf

+++ mojo/services/public/cpp/view_manager/lib/view_manager_synchronizer.cc
@@ -365,7 +365,8 @@ class SetViewContentsTransaction : public ViewManagerTransaction {
     memcpy(memory, &data[0], data.size());
 
     AllocationScope scope;
-    service()->SetViewContents(view_id_, duped.Pass(), data.size(),
+    service()->SetViewContents(view_id_, duped.Pass(),
+                               static_cast<uint32_t>(data.size()),
                                ActionCompletedCallback());
   }
   virtual void DoActionCompleted(bool success) OVERRIDE {
